premises of the prompt management
